New Movement code 50% done

+++ src/game/level/EntityComparator.java
@@ -1,7 +1,7 @@
 package game.level;
 
 import game.entity.Entity;
-import game.math.Vector2d;
+import game.math.Vector2i;
 
 import java.util.Comparator;
 
@@ -15,8 +15,8 @@ public class EntityComparator implements Comparator<Entity> {
 	
 	@Override
 	public int compare(Entity e0, Entity e1) {
-		Vector2d e0Pos = e0.getPos();
-		Vector2d e1Pos = e1.getPos();
+		Vector2i e0Pos = e0.getPos();
+		Vector2i e1Pos = e1.getPos();
 		if (e0Pos.y < e1Pos.y)
 			return -1;
 		if (e0Pos.y > e1Pos.y)
